src/app/shared/nav-bar/nav-bar.component.ts
import { AuthService } from './../../services/auth.service';
import {
  Component,
  ChangeDetectionStrategy,
  OnInit,
  OnDestroy
} from '@angular/core';
import { BreakpointObserver } from '@angular/cdk/layout';
import { Subscription } from 'rxjs';

@Component({
  selector: 'app-nav-bar',
  templateUrl: './nav-bar.component.html',
  styleUrls: ['./nav-bar.component.css'],
  changeDetection: ChangeDetectionStrategy.OnPush
})
export class NavBarComponent implements OnInit, OnDestroy {
  $subs: Subscription;
  isLoggedIn = false;

  constructor(
    // eslint-disable-next-line @typescript-eslint/no-unused-vars-experimental
    private breakpointObserver: BreakpointObserver,
    private authService: AuthService
  ) {}

  ngOnInit(): void {
    this.$subs = this.authService.observe().subscribe(isLoggedIn => {
      this.isLoggedIn = isLoggedIn;
    });
  }

  ngOnDestroy(): void {
    this.$subs.unsubscribe();
  }

  logout(): void {
    this.authService.logout();
  }
}
